+++ build.gradle
@@ -1,5 +1,5 @@
 plugins {
-    id 'com.github.johnrengelman.shadow' version '7.1.2'
+    id 'com.github.johnrengelman.shadow' version '5.1.0'
     id 'java'
     id 'application'
 }
@@ -29,7 +29,7 @@ dependencies {
     implementation group: 'org.openjfx', name: 'javafx-graphics', version: javaFxVersion, classifier: 'linux'
 }
 
-tasks.withType(Test) {
+run {
     standardInput = System.in
     enableAssertions = true
 }
