+++ src/main/java/Duke/Duke.java
@@ -1,5 +1,12 @@
 package Duke;
 import java.util.Scanner;
+import Duke.Data.Storage;
+import Duke.Exception.DukeException;
+import Duke.GUI.DialogBox;
+import Duke.GUI.GuiUi;
+import Duke.Handler.Parser;
+import Duke.Handler.Ui;
+import Duke.Tasks.TaskList;
 import javafx.scene.Scene;
 import javafx.scene.control.ScrollPane;
 import javafx.scene.control.TextField;
@@ -11,7 +18,6 @@ import javafx.application.Application;
 import javafx.scene.control.Button;
 import javafx.scene.control.Label;
 import javafx.scene.image.Image;
-import javafx.scene.image.ImageView;
 
 /**
  * Duke is the main class that will save and run the program
@@ -108,11 +114,19 @@ public class Duke extends Application {
 
         //functionality
         button.setOnMouseClicked((event) -> {
-            handleUserInput();
+            try {
+                handleUserInput();
+            } catch (DukeException e) {
+                e.printStackTrace();
+            }
         });
 
         userInput.setOnAction((event) -> {
-            handleUserInput();
+            try {
+                handleUserInput();
+            } catch (DukeException e) {
+                e.printStackTrace();
+            }
         });
         dialogContainer.heightProperty().addListener((observable) -> scrollPane.setVvalue(1.0));
     }
@@ -170,7 +184,7 @@ public class Duke extends Application {
      * Creates two dialog boxes, one echoing user input and the other containing Duke's reply and then appends them to
      * the dialog container. Clears the user input after processing.
      */
-    private void handleUserInput() {
+    private void handleUserInput() throws DukeException {
         Label userText = new Label(userInput.getText());
         Label dukeText = new Label(getResponse(userInput.getText()));
         dialogContainer.getChildren().addAll(
@@ -180,12 +194,8 @@ public class Duke extends Application {
         userInput.clear();
     }
 
-    public String getResponse(String input) {
-        //String response = ui.
+    public String getResponse(String input) throws DukeException {
         String response = guiUi.enterText();
-
-        trimUserInput(input);
-
         if (input.equals("list")) {
             response = guiUi.displayTask();
         } else if (input.startsWith("mark")) {
