+++ src/main/java/Duke/GUI/DialogBox.java
@@ -1,7 +1,6 @@
-package Duke;
+package Duke.GUI;
 import java.io.IOException;
 import java.util.Collections;
-
 import javafx.collections.FXCollections;
 import javafx.collections.ObservableList;
 import javafx.fxml.FXML;
