+++ src/main/java/Duke/Handler/Ui.java
@@ -1,4 +1,6 @@
-package Duke;
+package Duke.Handler;
+import Duke.Data.Storage;
+import Duke.Tasks.Task;
 import java.util.ArrayList;
 
 /**
